+++ common/routes/Channel/containers/ChannelContainer.js
@@ -9,7 +9,7 @@ import { Header } from '../../../components/Header';
 import Pagination from '../../../components/Pagination';
 import { ChannelSelecter, BlockItems } from '../components';
 import { Container, Loading, Margin10, NotFound } from '../../../components/Layout';
-import { DFP, OneAdICIP } from '../../../components/Ad';
+import { DFP } from '../../../components/Ad';
 import Helmet from 'react-helmet';
 
 const redial = {
@@ -36,7 +36,6 @@ const ChannelContainer = ({ channelPage, local, loadChannelData, menus, marquee 
   return (
     <Container>
       <Header menus={menus} marquee={marquee} />
-      <OneAdICIP />
       {isLoading && <Loading />}
       <ChannelSelecter channels={channels} selectedChannel={channel} loadChannelData={loadChannelData} />
       {!isLoading && newsList.length === 0 && <NotFound />}