+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changed
 - 依照業務需求修改 ads.txt @wb
+- 對調所有列表頁 DFP 與 OneAdICIP 的位置 @esbb48
 
 ## 1.0.46 - 2017-12-19
 ### Changed